Moved Query object to types.d.ts
