+++ frontend/src/js/types.d.ts
@@ -31,3 +31,24 @@ interface User {
     token: string
     groups: string[]
 }
+
+type Query = Partial<{
+    title: string
+    fileName: string
+    copyright: string
+    years: {
+        from?: number
+        to?: number
+    }
+    filesize: {
+        from?: number
+        to?: number
+    }
+    reviewed: {
+        from?: string
+        to?: string
+    }
+    active: 'all' | 'active' | 'inactive'
+    duplicatable: 'all' | 'duplicatable' | 'nonduplicatable'
+    metadata: Record<number,Metadata[]>
+}>
