+++ frontend/src/js/urls.ts
@@ -1,6 +1,5 @@
 //All hardcoded end-points live in this file
 const APP_URLS = {
-    CONTENT_LIST: `/api/content/`,
     LOGOUT: `/accounts/logout/`,
     METADATA: (id: number) => `/api/metadata/${id}/`,
     //api endpoint `/api/metadata_types/{type_id}/metadata/ ` returns all 
@@ -10,6 +9,7 @@ const APP_URLS = {
     METADATA_LIST: `/api/metadata/`,
     METADATA_TYPES: `/api/metadata_types/`,
     METADATA_TYPE: (id: number) => `/api/metadata_types/${id}/`,
+    CONTENT_LIST: `/api/content/`,
     TOAST_MESSAGE: ``,
     USER_INFO: `/api/get_user/`,
 }
